#118 - add review edition
